requirements.txt
requests==2.2.0
requests-oauthlib==0.4.0
simplejson==3.3.2
wsgiref==0.1.2
oauthlib==0.6.0
python-twitter==1.1


